# openapi.yaml
openapi: 3.0.0
info:
  version: '1.1'
  title: Ruokavälitys API
  description: API for both, managing the RV system and it's inventory, and browsing and purchasing products.
...
servers:
  - url: https://rv-backend-dev.herokuapp.com/
    description: Staging environment
  - url: https://rv-backend.herokuapp.com/
    description: Production environment
components:
  securitySchemes:
    normalAccessToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description:
        A JWT access token used for authenticating with the \"end user facing\" portition of the API.
        Can be obtained with any valid credentials from route `/api/v1/authenticate`.
    adminAccessToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description:
        A JWT access token used for authenticating with the administrations portition of the API.
        Can be obtained with administrator credentials from route `/api/v1/admin/authenticate`.
  schemas:
...
    Product:
      type: object
      required: 
        - barcode
        - name
        - category
        - weight
        - sellPrice
        - stock
      properties: 
        barcode: 
          type: string
          example: "6430039222212"
        name: 
          type: string
          example: Kombucha bakteerin makuinen bakteerijuoma
        category: 
          '$ref': '#/components/schemas/Category'
        weight: 
          type: integer
          example: 370
        sellPrice: 
          type: number
          example: 3.10
        stock: 
          type: integer
          example: 10
...
    Box:
      type: object
      required: 
        - boxBarcode
        - itemsPerBox
        - product
      properties: 
        boxBarcode:
          example: "01880335"
          type: string
        itemsPerBox: 
          type: integer
          example: 18
        product:
          '$ref': '#/components/schemas/Product'
...
    Category:
      type: object
      required: 
        - categoryId
        - description
      properties: 
        categoryId: 
          type: integer
          example: 10
        description: 
          type: string
          example: Drinks, soda with bottle return
...
  responses:
...
    NotAuthorizedErrorResponse:
      description: Your user lacks the required privilege level to perform this action.
      content:
        application/json:
          schema:
            type: object
            required:
              - error_code
              - message
            properties:
              error_code:
                type: string
                enum:
                  - not_authorized
              message:
                type: string
                example: Not authorized
    NotFoundErrorResponse:
      description: The resource your are requesting does not exist.
      content:
        application/json:
          schema:
            type: object
            required:
              - error_code
              - message
            properties:
              error_code:
                type: string
                enum:
                  - not_found
              message:
                type: string
                example: No product with barcode \'1312767817\' found
    BadRequestErrorResponse:
      description: The request you sent contains invalid values or is missing fields.
      content:
        application/json:
          schema:
            type: object
            required:
              - error_code
              - message
            properties:
              error_code:
                type: string
                enum:
                  - bad_request
                  - invalid_reference
              message:
                type: string
                example: Missing or invalid fields in request
              errors:
                type: array
                items:
                  type: string
                  example: 'boxBarcode should be a numeric 1-14 digit barcode' 
    IdentifierTakenErrorResponse:
      description: Creating a new resource failed, because a resource with the given identifier already exists.
      content:
        application/json:
          schema:
            type: object
            required:
              - error_code
              - message
            properties:
              error_code:
                type: string
                enum:
                  - identifier_taken
              message:
                type: string
                example: Barcode already in use.
paths: 
...
  /api/v1/admin/boxes: 
    get: 
      summary: Get a list of product boxes
      security:
        - adminAccessToken: []
      responses: 
        200: 
          description: Request fulfilled successfully
          content: 
            application/json: 
              schema: 
                type: object
                required: 
                  - boxes
                properties: 
                  boxes: 
                    type: array
                    items: 
                      '$ref': '#/components/schemas/Box'
        401:
          '$ref': '#/components/responses/NotAuthorizedErrorResponse'
    post: 
      summary: Create a new box of products
      security:
        - adminAccessToken: []
      responses: 
        201: 
          description: A new box created successfully
          content:
            application/json: 
              schema: 
                type: object
                required: 
                  - box
                properties: 
                  box: 
                    '$ref': '#/components/schemas/Box'
        401:
          '$ref': '#/components/responses/NotAuthorizedErrorResponse'
        400:
          '$ref': '#/components/responses/BadRequestErrorResponse'
        409:
          '$ref': '#/components/responses/IdentifierTakenErrorResponse'
      requestBody: 
        required: true
        content: 
          application/json: 
            schema: 
              type: object
              required: 
                - boxBarcode
                - itemsPerBox
                - productBarcode
              properties: 
                boxBarcode: 
                  type: string
                  example: '01766752'
                itemsPerBox: 
                  type: integer
                  example: 18
                productBarcode: 
                  type: string
                  example: '7314660003114'
  /api/v1/admin/boxes/{boxBarcode}: 
    get: 
      summary: Get the details of a specific product box
      security:
        - adminAccessToken: []
      parameters: 
        - name: boxBarcode
          required: true
          in: path
          schema: 
            type: string
      responses: 
        200: 
          description: Request fulfilled successfully
          content: 
            application/json: 
              schema: 
                type: object
                required: 
                  - box
                properties: 
                  box: 
                    '$ref': '#/components/schemas/Box'
        401:
          '$ref': '#/components/responses/NotAuthorizedErrorResponse'
        404:
          '$ref': '#/components/responses/NotFoundErrorResponse'
    patch: 
      summary: Update the details of a product box
      security:
        - adminAccessToken: []
      parameters: 
        - name: boxBarcode
          required: true
          in: path
          schema: 
            type: string
      responses: 
        200: 
          description: Box updated successfully
          content: 
            application/json: 
              schema: 
                type: object
                required: 
                  - box
                properties: 
                  box: 
                    '$ref': '#/components/schemas/Box'
        400:
          '$ref': '#/components/responses/BadRequestErrorResponse'
        401:
          '$ref': '#/components/responses/NotAuthorizedErrorResponse'
        404:
          '$ref': '#/components/responses/NotFoundErrorResponse'
      requestBody: 
        required: true
        content: 
          application/json: 
            schema: 
              type: object
              properties: 
                itemsPerBox: 
                  type: integer
                  example: 18
                productBarcode: 
                  type: string
                  example: '7314660003114'
    delete: 
      summary: Delete a product box
      security:
        - adminAccessToken: []
      parameters: 
        - name: boxBarcode
          required: true
          in: path
          schema: 
            type: string
      responses: 
        200: 
          description: Box deleted successfully
          content: 
            application/json: 
              schema: 
                type: object
                required: 
                  - deletedBox
                properties: 
                  deletedBox: 
                    '$ref': '#/components/schemas/Box'
        400:
          '$ref': '#/components/responses/BadRequestErrorResponse'
        401:
          '$ref': '#/components/responses/NotAuthorizedErrorResponse'
        404:
          '$ref': '#/components/responses/NotFoundErrorResponse'
...